HubspotCRM - Copy/requirements.txt
flask==2.2.5
flask-restx==1.0.5
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
flask-sqlalchemy==3.0.3
python-decouple==3.8
flask-migrate==4.0.5
Werkzeug==2.2.3